feat(highlight): allow extended highlighting patterns (#185) (#255)
Highlight the first capture group pattern, if there is a second
(typically nested) capture group, use it for keyword matching.
Allows highlighting `TODO(foo)` using `.*<((KEYWORDS)%(\(.{-1,}\))?):`.

Closes #185, closes #10

This is similar to #180, however that PR's matching behaviour is not
backwards compatible with the default highlight pattern used. This PR
retains the behaviour that only the first capturing group is
highlighted.