fix(config): select `fg` by maximizing contrast rather than check if a color is dark (#316)
## Description

This PR changes a bit of of the internal logic used for calculating the
foreground for a given background color. Currently it uses an algorithm
to guess if the dark or light color is better. This rather uses the
contrast ratio metric as described by the WCAG to maximize visual
difference between the foreground and background.

Let me know what you think of this solution! Thanks so much for a great
plugin!

EDIT: moved from Euclidian distance in the RGB space to the official
contrast definition by the WCAG for identifying the contrast between
foreground and background colors which accounts for difference in color
perception to the human eye.
(https://www.w3.org/WAI/WCAG21/Understanding/contrast-minimum.html)

## Related Issue(s)

Fixes #315 

## Screenshots

Here is a color that resolves incorrectly without this PR:


![image](https://github.com/user-attachments/assets/aba6bdaf-bd69-4f16-90bf-d8b7d4f130a1)

Here is the same color with this PR:


![image](https://github.com/user-attachments/assets/db7372a1-cba8-4f0d-bd00-0c0931e15dad)