feat: Allow showing TodoQuickFix list in sorted order
This is specially usefull when documenting  observations as
comments while working on a project in order to save the context for
future.

A keyword like PIN can be defined, with a search regexp as [[PIN \d{2}]]
and the comments can be numbered as # PIN: 00, # PIN: 01.. Then using
the TodoQuickFix command will show all the pins in the sorted order,
which makes it really easy to go over the past observations etc.